refactor(ai): military filter, matrix merge, s7d split (#3941)
Unify EXPAND/COLONIAL military destination partitioning behind a shared
region filter, collapse the gp-blocker peace matrix part shards into one
contract with case-table support modules, split seed42 S7-D feedstock
probes under test/support/s7d/, and add repo.ai_test_no_matrix_part_shards.

Refs #3941
